requirements.txt
streamlit
langchain
langgraph
langgraph-checkpoint-sqlite
sentence-transformers
faiss-cpu
python-dotenv
requests
openai
langchain-community
langchain-openai
langchain-huggingface
xmltodict
# Evaluation harness (optional — only needed to run eval/)
ragas
deepeval
datasets
